util: refactor all funcs in genRuleFile
